// packages/shared/src/generated/Profile.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Latch } from "./Latch";
import type { ProfileConfig } from "./ProfileConfig";

export type Profile = { id: string, name: string, color: string, createdAtMs: number, config: ProfileConfig, latch: Latch, };